+++ src/Exceptions/FontManagerNotInitializedException.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace WForest.Exceptions
+{
+    public class FontManagerNotInitializedException : Exception
+    {
+        public FontManagerNotInitializedException()
+        {
+        }
+
+        public FontManagerNotInitializedException(string message) : base(message)
+        {
+        }
+
+        public FontManagerNotInitializedException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}